AEAD: Move each key type into its own submodule.
Encapsulate each key type in its own module.
