+++ Cargo.toml
@@ -117,10 +117,13 @@ include = [
     "src/aead/gcm/gcm_nohw.rs",
     "src/aead/less_safe_key.rs",
     "src/aead/nonce.rs",
+    "src/aead/opening_key.rs",
     "src/aead/poly1305.rs",
     "src/aead/poly1305_test.txt",
     "src/aead/quic.rs",
+    "src/aead/sealing_key.rs",
     "src/aead/shift.rs",
+    "src/aead/unbound_key.rs",
     "src/agreement.rs",
     "src/arithmetic.rs",
     "src/arithmetic/bigint.rs",
